requirements.txt
requests
pandas
psycopg2-binary
SQLAlchemy
python-dotenv
kafka-python
cassandra-driver